move rmn proxy to v1_0
