+++ integration-tests/deployment/ccip/view/v1_0/rmn_proxy_contract.go
@@ -1,4 +1,4 @@
-package v1_6
+package v1_0
 
 import (
 	"fmt"
